// src/pages/login/__generated__/LoginUserMutation.graphql.ts
export type UserLoginInput = {
  email: string;
  password: string;
};
export type LoginUserMutation$variables = {
  input: UserLoginInput;
};
export type LoginUserMutation$data = {
  readonly userLogin: {
    readonly userAuth: {
      readonly person: {
        readonly email: string;
        readonly firstName: string;
        readonly lastName: string;
      } | null;
    } | null;
    readonly userErrors: ReadonlyArray<{
      readonly field: ReadonlyArray<string> | null;
      readonly message: string;
    } | null> | null;
    readonly userToken: string | null;
  } | null;
};
export type LoginUserMutation = {
  response: LoginUserMutation$data;
  variables: LoginUserMutation$variables;
};